requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
python-dotenv>=1.0.1
requests>=2.32.0
typing-extensions>=4.12.0
redis>=5.0.0
